+++ src/model/PlayModel.hpp
@@ -1,12 +1,8 @@
 #ifndef PLAY_MODEL_HPP
 #define PLAY_MODEL_HPP
 
-#include <utility>
 #include <stdint.h>
-#include <vector>
 
-#include "../model/Direction.hpp"
-#include "../model/TileType.hpp"
 #include "../model/State.hpp"
 #include "../model/BaseModel.hpp"
 
@@ -14,9 +10,9 @@ class PlayModel : public BaseModel {
 
 public:
     long getTimeClock();
-    void setTimeClock(uint16_t timeClockValue);
-    void decrementTimeClock(uint16_t step);
-    void incrementTimeClock(uint16_t step);
+    void setTimeClock(long timeClockValue);
+    void decrementTimeClock(long step);
+    void incrementTimeClock(long step);
 
     // Garbage
     static PlayModel* GetInstance();
